Mac OSX: Added code signing capability with --osx-codesign-identity arg.
OSX only: if --osx-codesign-identity is specified on the command-line,
then we will use the Xcode Developer Tools `codesign` utility to sign all
embedded binaries as well as the .app bundle itself. Note that this
option may also be given to PKG and BUNDLE for finer-grained control as
well. Defaults to off.